get_open_orders extracted
