+++ src/core/exchanges/main/mod.rs
@@ -2,3 +2,4 @@ pub mod cancel_order;
 pub mod create_order;
 pub mod exchange;
 pub mod features;
+pub mod get_open_orders;
